requirements.txt
# Core ML & Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
lightgbm>=3.3.0
xgboost>=1.5.0
imbalanced-learn>=0.9.0

# Model Interpretability
shap>=0.41.0
lime>=0.2.0

# Privacy & Security
cryptography>=36.0.0
hashlib-additional>=1.0.0

# Web Frameworks
streamlit>=1.25.0
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=20.1.0

# API & Serialization
pydantic>=1.10.0
fastapi>=0.95.0
uvicorn>=0.22.0
requests>=2.28.0

# Visualization
plotly>=5.14.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
python-dotenv>=0.20.0
joblib>=1.1.0
tqdm>=4.64.0

# Testing
pytest>=7.1.0
pytest-cov>=3.0.0
pytest-mock>=3.8.0

# Data Generation
faker>=18.0.0
scipy>=1.9.0

# Monitoring & Logging
loguru>=0.6.0

# Optional: Advanced features
# tensorflow>=2.10.0  # For deep learning models
# torch>=1.12.0  # Alternative DL framework
